[REF] hr_cae_promotion: separate module
