+++ hr_cae/models/__init__.py
@@ -3,4 +3,3 @@ from . import res_partner
 from . import hr_applicant
 from . import hr_employee
 from . import hr_exemption
-from . import hr_promotion
