+++ hr_cae/security/ir.model.access.csv
@@ -19,5 +19,3 @@ access_hr_sector_user,access_hr_sector,model_hr_sector,hr.group_hr_user,1,0,0,0
 access_hr_sector_manager,access_hr_sector,model_hr_sector,hr.group_hr_manager,1,1,1,1
 access_hr_contribution_exemption_user,access_hr_contribution_exemption,model_hr_contribution_exemption,hr.group_hr_user,1,0,0,0
 access_hr_contribution_exemption_manager,access_hr_contribution_exemption,model_hr_contribution_exemption,hr.group_hr_manager,1,1,1,1
-access_hr_promotion_user,access_hr_promotion,model_hr_promotion,hr.group_hr_user,1,0,0,0
-access_hr_promotion_manager,access_hr_promotion,model_hr_promotion,hr.group_hr_manager,1,1,1,1
