+++ hr_cae/views/hr_applicant.xml
@@ -63,9 +63,6 @@
                         <field name="salary_expected_extra" class="oe_inline" placeholder="Extra advantages..."/>
                     </div>
                 </group>
-                <group string="Promotion" name="promotion">
-                    <field name="promotion_id"/>
-                </group>
                 <group string="Professional Experience" name="professional_experience">
                     <field name="origin_status_id"/>
                     <field name="professional_experience"/>
