+++ hr_cae/views/hr_employee.xml
@@ -26,9 +26,6 @@
             </xpath>
 
             <xpath expr="//page[@name='public']//field[@name='job_id']/.." position="after">
-                <group name="Promotion">
-                    <field name="promotion_id"/>
-                </group>
                 <group name="Earnings">
                     <field name="turnover_minimum"/>
                 </group>
