+++ hr_cae_promotion/README.rst
@@ -1,0 +1,58 @@
+=================
+HR CAE Promotions
+=================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-odoo_cae%2Fodoo--addons--hr--incubator-lightgray.png?logo=github
+    :target: https://github.com/odoo_cae/odoo-addons-hr-incubator/tree/12.0/hr_cae_promotion
+    :alt: odoo_cae/odoo-addons-hr-incubator
+
+|badge1| |badge2| |badge3| 
+
+Add `hr.promotion` to handle groups of applicants in a CAE - Cooperative Activité Emploi
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/odoo_cae/odoo-addons-hr-incubator/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/odoo_cae/odoo-addons-hr-incubator/issues/new?body=module:%20hr_cae_promotion%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* Coop IT Easy SCRL
+
+Contributors
+~~~~~~~~~~~~
+
+* Manuel Claeys Bouuaert <manuel@coopiteasy.be>
+* Robin Keunen <robin@coopiteasy.be>
+
+Maintainers
+~~~~~~~~~~~
+
+This module is part of the `odoo_cae/odoo-addons-hr-incubator <https://github.com/odoo_cae/odoo-addons-hr-incubator/tree/12.0/hr_cae_promotion>`_ project on GitHub.
+
+You are welcome to contribute.
