+++ hr_cae_promotion/__manifest__.py
@@ -3,30 +3,21 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
 {
-    "name": "HR CAE",
-    "summary": "Employee HR in a CAE - Cooperative Activité Emploi",
+    "name": "HR CAE Promotions",
+    "summary": "Add promotions as groups of employees.",
     "author": "Coop IT Easy SCRL",
     "website": "https://coopiteasy.be",
     "category": "Human Resources",
     "version": "12.0.1.0.0",
     "license": "AGPL-3",
-    "depends": [
-        "account_banking_sepa_direct_debit",
-        "base_location_geonames_import",
-        "hr",
-        "hr_recruitment",
-        "hr_employee_age",
-        "l10n_fr_department",
-    ],
+    "depends": ["hr_cae"],
     "data": [
         "security/ir.model.access.csv",
-        "views/hr.xml",
         "views/hr_applicant.xml",
         "views/hr_employee.xml",
-        "views/hr_exemption.xml",
-        "views/res_partner.xml",
-        "data/data.xml",
+        "views/hr_promotion.xml",
     ],
+    "demo": ["demo/demo.xml"],
     "installable": True,
     "application": False,
 }
