+++ hr_cae_promotion/demo/demo.xml
