+++ hr_cae_promotion/models/__init__.py
@@ -1,0 +1,3 @@
+from . import hr_applicant
+from . import hr_employee
+from . import hr_promotion
