+++ hr_cae_promotion/models/hr_employee.py
@@ -1,0 +1,14 @@
+# Copyright 2019 Coop IT Easy SCRL fs
+#   Manuel Claeys Bouuaert <manuel@coopiteasy.be>
+#   Robin Keunen <robin@coopiteasy.be>
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
+
+from odoo import fields, models
+
+
+class Employee(models.Model):
+    _inherit = "hr.employee"
+
+    promotion_id = fields.Many2one(
+        comodel_name="hr.promotion", string="Promotion", required=False
+    )
