+++ hr_cae_promotion/readme/CONTRIBUTORS.rst
@@ -1,0 +1,2 @@
+* Manuel Claeys Bouuaert <manuel@coopiteasy.be>
+* Robin Keunen <robin@coopiteasy.be>
