+++ hr_cae_promotion/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+Add `hr.promotion` to handle groups of applicants in a CAE - Cooperative Activité Emploi
