+++ hr_cae_promotion/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_hr_promotion_user,access_hr_promotion,model_hr_promotion,hr.group_hr_user,1,0,0,0
+access_hr_promotion_manager,access_hr_promotion,model_hr_promotion,hr.group_hr_manager,1,1,1,1
