+++ hr_cae_promotion/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_hr_cae_promotion
