+++ hr_cae_promotion/views/hr_applicant.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+     Copyright 2019 Coop IT Easy
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+<odoo>
+
+    <record id="view_applicant_form" model="ir.ui.view">
+        <field name="name">hr.applicant.form</field>
+        <field name="model">hr.applicant</field>
+        <field name="inherit_id" ref="hr_recruitment.crm_case_form_view_job"/>
+        <field name="arch" type="xml">
+
+            <!-- group Job Position -->
+            <xpath expr="//group[@name='job_position']" position="after">
+                <group string="Promotion" name="promotion">
+                    <field name="promotion_id"/>
+                </group>
+            </xpath>
+
+
+        </field>
+    </record>
+</odoo>
