+++ hr_cae_promotion/views/hr_employee.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+     Copyright 2019 Coop IT Easy
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+<odoo>
+
+    <record id="view_employee_form" model="ir.ui.view">
+        <field name="name">hr.employee.form</field>
+        <field name="model">hr.employee</field>
+        <field name="inherit_id" ref="hr.view_employee_form"/>
+        <field name="arch" type="xml">
+
+            <!-- page work information -->
+            <xpath expr="//group[@name='Earnings']" position="before">
+                <group name="Promotion">
+                    <field name="promotion_id"/>
+                </group>
+            </xpath>
+
+        </field>
+    </record>
+</odoo>
